+++ test/watch/src/multi-format.ts
@@ -1,1 +1,0 @@
-export const multiFormat = "watch multi format"